+++ packages/vite-plugin-kitbook/src/constants.ts
@@ -4,3 +4,5 @@ export default augmentSvelteConfigForKitbook(config)`;
 
 export const VIRTUAL_MODULES_IMPORT_ID = 'virtual:kitbook-modules';
 export const RESOLVED_VIRTUAL_MODULES_IMPORT_ID = '\0' + VIRTUAL_MODULES_IMPORT_ID;
+
+export const DEFAULT_IMPORT_MODULE_GLOBS = ['/src/**/*.{md,svx,svelte,variants.ts}', '/README.md'];