+++ packages/vite-plugin-kitbook/src/virtual/output/importModules.js
@@ -1,6 +1,6 @@
 import { groupColocatedModulesIntoPages, pagesStore } from "kitbook";
-const modules = import.meta.glob(["/src/**/*.{md,svx,svelte,variants.ts}", "/README.md"]);
-const rawModules = import.meta.glob(["/src/**/*.{md,svx,svelte,variants.ts}", "/README.md"], { as: "raw" });
+const modules = import.meta.glob(["REPLACE_WITH_MODULE_GLOBS"]);
+const rawModules = import.meta.glob(["REPLACE_WITH_MODULE_GLOBS"], { as: "raw" });
 export const pages = groupColocatedModulesIntoPages(modules, rawModules);
 const WrapRootLayoutMap = import.meta.glob(["/src/.kitbook/WrapRootLayout.svelte"], { eager: true, import: "default" });
 export const WrapRootLayout = WrapRootLayoutMap["/src/.kitbook/WrapRootLayout.svelte"];